requirements.txt
anthropic>=0.40.0
fastapi>=0.110.0
google-search-results>=2.4.2
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.29.0
